refactor(api): move service logic to separate package
