+++ api/service/create.go
@@ -1,0 +1,131 @@
+// Copyright (c) 2023 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+package service
+
+import (
+	"fmt"
+	"net/http"
+	"time"
+
+	"github.com/go-vela/server/router/middleware/org"
+	"github.com/go-vela/server/router/middleware/user"
+
+	"github.com/go-vela/server/database"
+	"github.com/go-vela/server/router/middleware/build"
+	"github.com/go-vela/server/router/middleware/repo"
+	"github.com/go-vela/server/util"
+
+	"github.com/go-vela/types/constants"
+	"github.com/go-vela/types/library"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+)
+
+// swagger:operation POST /api/v1/repos/{org}/{repo}/builds/{build}/services services CreateService
+//
+// Create a service for a build in the configured backend
+//
+// ---
+// produces:
+// - application/json
+// parameters:
+// - in: path
+//   name: org
+//   description: Name of the org
+//   required: true
+//   type: string
+// - in: path
+//   name: repo
+//   description: Name of the repo
+//   required: true
+//   type: string
+// - in: path
+//   name: build
+//   description: Build number
+//   required: true
+//   type: integer
+// - in: body
+//   name: body
+//   description: Payload containing the service to create
+//   required: true
+//   schema:
+//     "$ref": "#/definitions/Service"
+// security:
+//   - ApiKeyAuth: []
+// responses:
+//   '201':
+//     description: Successfully created the service
+//     schema:
+//       "$ref": "#/definitions/Service"
+//   '400':
+//     description: Unable to create the service
+//     schema:
+//       "$ref": "#/definitions/Error"
+//   '500':
+//     description: Unable to create the service
+//     schema:
+//       "$ref": "#/definitions/Error"
+
+// CreateService represents the API handler to create
+// a service for a build in the configured backend.
+func CreateService(c *gin.Context) {
+	// capture middleware values
+	b := build.Retrieve(c)
+	o := org.Retrieve(c)
+	r := repo.Retrieve(c)
+	u := user.Retrieve(c)
+
+	entry := fmt.Sprintf("%s/%d", r.GetFullName(), b.GetNumber())
+
+	// update engine logger with API metadata
+	//
+	// https://pkg.go.dev/github.com/sirupsen/logrus?tab=doc#Entry.WithFields
+	logrus.WithFields(logrus.Fields{
+		"build": b.GetNumber(),
+		"org":   o,
+		"repo":  r.GetName(),
+		"user":  u.GetName(),
+	}).Infof("creating new service for build %s", entry)
+
+	// capture body from API request
+	input := new(library.Service)
+
+	err := c.Bind(input)
+	if err != nil {
+		retErr := fmt.Errorf("unable to decode JSON for new service for build %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusBadRequest, retErr)
+
+		return
+	}
+
+	// update fields in service object
+	input.SetRepoID(r.GetID())
+	input.SetBuildID(b.GetID())
+
+	if len(input.GetStatus()) == 0 {
+		input.SetStatus(constants.StatusPending)
+	}
+
+	if input.GetCreated() == 0 {
+		input.SetCreated(time.Now().UTC().Unix())
+	}
+
+	// send API call to create the service
+	err = database.FromContext(c).CreateService(input)
+	if err != nil {
+		retErr := fmt.Errorf("unable to create service for build %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusInternalServerError, retErr)
+
+		return
+	}
+
+	// send API call to capture the created service
+	s, _ := database.FromContext(c).GetServiceForBuild(b, input.GetNumber())
+
+	c.JSON(http.StatusCreated, s)
+}
