+++ api/service/update.go
@@ -1,0 +1,152 @@
+// Copyright (c) 2023 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+package service
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/go-vela/server/router/middleware/org"
+	"github.com/go-vela/server/router/middleware/user"
+
+	"github.com/go-vela/server/database"
+	"github.com/go-vela/server/router/middleware/build"
+	"github.com/go-vela/server/router/middleware/repo"
+	"github.com/go-vela/server/router/middleware/service"
+	"github.com/go-vela/server/util"
+
+	"github.com/go-vela/types/library"
+
+	"github.com/gin-gonic/gin"
+	"github.com/sirupsen/logrus"
+)
+
+//
+// swagger:operation PUT /api/v1/repos/{org}/{repo}/builds/{build}/services/{service} services UpdateService
+//
+// Update a service for a build in the configured backend
+//
+// ---
+// produces:
+// - application/json
+// parameters:
+// - in: path
+//   name: org
+//   description: Name of the org
+//   required: true
+//   type: string
+// - in: path
+//   name: repo
+//   description: Name of the repo
+//   required: true
+//   type: string
+// - in: path
+//   name: build
+//   description: Build number
+//   required: true
+//   type: integer
+// - in: path
+//   name: service
+//   description: Service number
+//   required: true
+//   type: integer
+// - in: body
+//   name: body
+//   description: Payload containing the service to update
+//   required: true
+//   schema:
+//     "$ref": "#/definitions/Service"
+// security:
+//   - ApiKeyAuth: []
+// responses:
+//   '200':
+//     description: Successfully updated the service
+//     schema:
+//       "$ref": "#/definitions/Service"
+//   '400':
+//     description: Unable to update the service
+//     schema:
+//       "$ref": "#/definitions/Error"
+//   '500':
+//     description: Unable to update the service
+//     schema:
+//       "$ref": "#/definitions/Error"
+
+// UpdateService represents the API handler to update
+// a service for a build in the configured backend.
+func UpdateService(c *gin.Context) {
+	// capture middleware values
+	b := build.Retrieve(c)
+	o := org.Retrieve(c)
+	r := repo.Retrieve(c)
+	s := service.Retrieve(c)
+	u := user.Retrieve(c)
+
+	entry := fmt.Sprintf("%s/%d/%d", r.GetFullName(), b.GetNumber(), s.GetNumber())
+
+	// update engine logger with API metadata
+	//
+	// https://pkg.go.dev/github.com/sirupsen/logrus?tab=doc#Entry.WithFields
+	logrus.WithFields(logrus.Fields{
+		"build":   b.GetNumber(),
+		"org":     o,
+		"repo":    r.GetName(),
+		"service": s.GetNumber(),
+		"user":    u.GetName(),
+	}).Infof("updating service %s", entry)
+
+	// capture body from API request
+	input := new(library.Service)
+
+	err := c.Bind(input)
+	if err != nil {
+		retErr := fmt.Errorf("unable to decode JSON for service %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusBadRequest, retErr)
+
+		return
+	}
+
+	// update service fields if provided
+	if len(input.GetStatus()) > 0 {
+		// update status if set
+		s.SetStatus(input.GetStatus())
+	}
+
+	if len(input.GetError()) > 0 {
+		// update error if set
+		s.SetError(input.GetError())
+	}
+
+	if input.GetExitCode() > 0 {
+		// update exit_code if set
+		s.SetExitCode(input.GetExitCode())
+	}
+
+	if input.GetStarted() > 0 {
+		// update started if set
+		s.SetStarted(input.GetStarted())
+	}
+
+	if input.GetFinished() > 0 {
+		// update finished if set
+		s.SetFinished(input.GetFinished())
+	}
+
+	// send API call to update the service
+	err = database.FromContext(c).UpdateService(s)
+	if err != nil {
+		retErr := fmt.Errorf("unable to update service %s: %w", entry, err)
+
+		util.HandleError(c, http.StatusInternalServerError, retErr)
+
+		return
+	}
+
+	// send API call to capture the updated service
+	s, _ = database.FromContext(c).GetServiceForBuild(b, s.GetNumber())
+
+	c.JSON(http.StatusOK, s)
+}
